pyproject.toml
[project]
name = "mockapi"
version = "0.0.1"
description = ""
authors = [
    {name = "Aleksei Devyatkin",email = "aleksei@devyatkin.dev"}
]
license = "LICENSE"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
preview = true # (E271,E203) https://docs.astral.sh/ruff/preview/

unfixable = [
    "ERA", # do not autoremove commented out code
]

exclude = [
    ".git",
    "__pycache__",
    "devops",
]

ignore = [
    "A001", # Variable {name} is shadowing a Python builtin
    "A002", # Argument {name} is shadowing a Python builtin
    "A003", # Class attribute {name} is shadowing a Python builtin
    "B012", # {name} inside finally blocks cause exceptions to be silenced
    "E501", # Line too long ({width} > {limit} characters)
    "B904", # raise-without-from-inside-except
    "E203", # Whitespace before '{symbol}'
    "F841", # assigned to but never used

]

select = [
    "A", # same flake8-builtins
    "B", # same flake8-bugbear
    "E", # pycodestyle
    "F", # pyflakes
    "Q", # same flake8-quotes
    "W", # pycodestyle
    "E271", # Multiple spaces after keyword
    "ERA", # flake8-eradicate/eradicate
    "I", # same isort
    "T20",  # flake8-print
]

# Same as Black.(for future format in ruff)
line-length = 120

# Assume Python 3.7
target-version = "py37"

[tool.ruff.isort]
known-first-party = ["api", "core"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
